// src/api/models/album-preview.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AlbumPreview
 */
export interface AlbumPreview {
    /**
     * 앨범 ID
     * @type {string}
     * @memberof AlbumPreview
     */
    'id': string;
    /**
     * 앨범 제목
     * @type {string}
     * @memberof AlbumPreview
     */
    'title': string;
    /**
     * 앨범 커버 이미지 URL
     * @type {string}
     * @memberof AlbumPreview
     */
    'imageUrl': string;
    /**
     * 아티스트 목록
     * @type {Array<string>}
     * @memberof AlbumPreview
     */
    'artists': Array<string>;
    /**
     * 평균 평점
     * @type {number}
     * @memberof AlbumPreview
     */
    'averageRating': number;
    /**
     * 평가 개수
     * @type {number}
     * @memberof AlbumPreview
     */
    'ratingCount': number;
    /**
     * 현재 사용자가 준 평점 (로그인 시에만)
     * @type {number}
     * @memberof AlbumPreview
     */
    'userRating'?: number;
    /**
     * 현재 사용자가 평가했는지 여부 (로그인 시에만)
     * @type {boolean}
     * @memberof AlbumPreview
     */
    'isRated'?: boolean;
}